requirements.txt
pandas==2.0.3
pyyaml==6.0.1
numpy==1.24.3
